Make profile properties respect their contract
The profile module specifies a contract that each profile property must
respect. This change makes all properties respect the contract with
regard to definedness, default values and hierarchy.

To ease upholding certain aspects of the contract the new utility method
all_properties() was added.

The documentation of resolver.source4 and resolver.source6 is fixed to
not state that undef is a valid value. This brings the property
documentation in line with both implementation and the contractual
requirements for properties.